requirements_dev.txt
pytest==8.4.1
pytest-django==4.9.0